% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchGlobe.R
\name{searchGlobe}
\alias{searchGlobe}
\title{Search terms globally}
\usage{
searchGlobe(terms, filename)
}
\arguments{
\item{terms}{List of terms to search}

\item{filename}{name prefix for excel output}
}
\value{
An excel sheet named filename_US.xlsx with sheets interest_over_time, interest_by_country, interest_by_dma, interest_by_city, related_topics, related_queries
}
\description{
Search terms globally
}
\examples{
searchGlobe(c("Trump", "Biden"), "politics" )
}
