% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R, R/searchUS.R
\name{searchUS}
\alias{searchUS}
\title{Add together two numbers}
\usage{
searchUS(terms, filename)

searchUS(terms, filename)
}
\arguments{
\item{terms}{List of terms to search}

\item{filename}{name prefix for excel output}
}
\value{
An excel sheet named filename_US.xlsx with sheets interest_over_time, interest_by_region, interest_by_dma, interest_by_city, related_topics, related_queries

An excel sheet named filename_US.xlsx with sheets interest_over_time, interest_by_region, interest_by_dma, interest_by_city, related_topics, related_queries
}
\description{
Add together two numbers

Search terms USA
}
\examples{
search(US, c("Trump", "Biden"), "politics" )
add(10, 1)
searchUS(c("Trump", "Biden"), "politics" )
}
