% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchUS.R
\name{searchUS}
\alias{searchUS}
\title{Search terms USA}
\usage{
searchUS(terms, filename)
}
\arguments{
\item{terms}{List of terms to search}

\item{filename}{name prefix for excel output}
}
\value{
An excel sheet named filename_US.xlsx with sheets interest_over_time, interest_by_region, interest_by_dma, interest_by_city, related_topics, related_queries
}
\description{
Search terms USA
}
\examples{
searchUS(c("Trump", "Biden"), "politics" )
}
